don't assume gOut["content"] is variant editing
In public_html/lib/template.php, gOut["content"] was assumed to contain only a
variant-editing page, and so this html was wrapped with a form that sent it to
submit.php. Now gOut["content"] is used for other types of content, and it
breaks if one tries to include a form within this new content.

To fix, the form html addition has been moved so it is only added to the
variant display content (which is in public_html/index.html).