Add private_genomes table, store info from uploads
Adding "private_genomes" table to "evidence" database in MySQL to store data
pertaining to uploaded genomes. (This is done in "evidence.php") The new table
includes openID of user that uploaded, and the shasum used to uniquely identify
the file (under which it is stored).

Modifying "template.php" and "genome_upload.php" to collect this information
and store it in "private_genomes".