Add export and scrub of eb_users table
Adding export of the users table, which includes oid, name, and email.
Because we don't want to be sharing email addresses, we want to scrub
this exported table so the name is retained but the email is hidden,
so this column is scrubbed from the dump using command line perl.