Clean up and update installation
A lot of changes to clean up installation and update instructions.
Mostly these reflect the following motivations:

- Remove usage of Trait-o-matic's SQL databases, these are no longer
used for genome processing. The only SQL we use now is the GET-Evidence
SQL data, some of which is redundant with creating Trait-o-matic's DB.
Changes include:
  - config.default.py: delete username/password for db logins
  - gff_gete-evidence_map.py, gff_nonsynonymous_filter_refflact_file.py:
    remove unnecessary import of db username/password
  - install-user.sh: remove unnecessary db-relevant commands
  - setup-db.sh: delete
  - setup-external-data.sh: remove most download and import to db

- Remove any running of scripts as root using sudo (but still retain
running install-user.sh under www-data using sudo). This is because the
scripts can be mucking around in system stuff without this being clear
to the user, e.g. "Do not wipe /tmp on reboot" (!)
Changes include:
  - install-sysconfig.sh: delete, move anything where root is
    needed to INSTALL instructions.
  - php-ini-update.txt: deleted, was used by install-sysconfig.sh to
    automatically change /etc/php5/apache2/php.ini