More bug fixes
The old code assumed there would be one like for allele 1, then one for
allele 2. It turns out there are some instances of multiple lines for one
or both alleles, so the code now reads as much as it needs to.

This also requires a "buffer", so it is always checking one line ahead to
so it can check that it has read in all of allele 2.