Allow GFF files to have end before start
This is breaking the spec. We're doing this so we can unambiguously
refer to an insertion by position alone. According to GFF spec, a
single base substitution starts and ends at the same position.
To distinguish the insertion as having a different position,
we describe it as having an "end" one base before "start".