Interpret indels and larger substitutions
A large rewrite of the nonsynonymous caller. Instead of finding the
frame, pulling the correct three bases, then finding the amino acid
for the codon, instead both "reference" and "variant" coding
sequences are inferred, translated, and differences between these
are reported.

As part of this, there is now a sanity check for reference coding
regions as being correctly formed -- they should have a multiple of
three, no stop codons in the middle, and a stop codon at the end.