Use multiprocessing for genome analysis subprocess
This introduces usage of python's multiprocessing module. In the
future we would like to have a queue which limits the number of
processes running at any one time. Unfortunately, this version
doesn't queue -- it simply allows many to run simultaneously. Notably
this is what the old version did as well (which started shell script
subprocesses), so presumably this is no worse in that respect.

I'm making change in how subprocesses are created so I can reduce
usage of shell script in the server. Previously, the entire genome
analysis was contained within a very long shell script (38 lines!).