Convert gff_getevidence_map.py to use a flatfile
'gff_getevidence_map.py': This previously used MySQL queries to look
for GET-Evidence hits. This was the only remaining module used by
the genome analysis server that used MySQL - this was blocking us from
sending genome processing jobs to the cloud. It now uses a
JSON-formatted flat file containing GET-Evidence data. In addition
it can now be called as a GFF-string generator, allowing it to be
added to our generator chaining.

'download.php': A function is added to generate the JSON-formatted
file that gff_getevidence_map.py needs.

'Makefile' was updated to generate the file using 'download.php' when
'make daily' is run.

'INSTALL': Instructions for setting up the 'make daily' cron job and
an initial run of it is added.

'UPGRADE': Instruction to run 'make daily' is added.

'.gitignore': A line to ignore the file made is added.

'config.default.py': A name referring to the file is added.

'trait-o-matic-server.py': Updated to correctly call the new
gff_getevidence_map.py.

This commit fixes #521 and references #426
Because latest-flat.tsv wasn't used, it also closes #520