Use get_metadata.header to just look at top of GFF
We'd like to determine (1) the build used as reference genome for a GFF and
(2) whether a file has "REF" lines (regions called as matching reference)
before genome processing. (1) is important because it determines which
twobit, dbSNP, and transcript files we should be using. (2) is important
because we should only try to use "gff_call_uncovered" if the GFF file has
regions called as matching reference.

Because this only needs us to examine the header (and a couple more lines,
if looking for "REF" lines), we can use get_metadata.header to do it quickly
rather than scanning the whole GFF file.