Use subprocess to pipe conversion and sort
Not a single intermediate file saved from the processing now. Instead of
sorting to file, the sorted data is piped to the processing generators.
In addition, if necessary conversion is done to the input and piped to
sort.

Opening the source data to infer format type and closing it is causing
an error when the pipe is closed. This is showing up in the log, but I
don't know how to avoid it. For example: "bzcat: I/O or other error,
bailing out."