Add 23andme interpretation
23andme_to_gff.py -> gff_from_23andme.py
Updated the conversion script and moved it to a standard python name
(names may not start with a number).

genome_analyzer.py is modified to detect 23andme data and call the
conversion script.

gff_twobit_query.py is modified to detect when "alleles" matches the
reference allele. When this is true, the record "feature" is set to
"REF" and the "attributes" field is made empty. (23andme data contains
positions matching reference. These cannot be recognized as such during
conversion -- it has to happen at this step when the reference allele
is checked.)

Note: we don't currently accept zip-compressed files in the upload,
although it's the default compression for 23andme data. The file
should be uncompressed, gzip, or bzip2.