config.py permissions issues no longer necessary
As I understand it, the reason config.py had this complicated copying,
editing, and permissions assignment was because it contained the
password needed to access the MySQL database stuff. Now that this is
totally removed (all analysis is done with flat files), we can just
use the file git is tracking.