Add "comp to hyp" for eval benign to calculators
Adding a "compare to hypothetical variant" test, for evaluating the
significance of observations when attempted to assign significance
to a "benign/nonpathogenic" hypothesis. The allele frequency for some
disease-causing hypothetical variant is estimated based on given
disease prevalence, percent-explained, penetrance, and inheritance.
Observations are then tested to see if they are significantly higher
than this hypothesized allele's frequency.

In addition, some explanation is added to the case/control and
penetrance analysis to describe in what contexts this calculator should
be used.