Add 2 second timeout to fishersexact.php
Fisher's Exact is doing a lot of factorial calculations and looping --
sometimes the numbers are just too big for it. Although the
"n choose k" calculation gets tough for big numbers, it's not merely
that -- another problem that the two-tailed test requires you to test
all other possible tables which end up with the same sum values (for
the a, b, c, d, all other versions where a+b, a+c, b+c, and c+d are the
same) -- it's the combination of these tasks that gets time consuming.

Rather than try to figure out the right limits to set on sample size
and/or how often the loop should go, I added a time check within it &
set it to abort after 2 seconds.