Fix breaks introduced with options=dict() default
The options produced by optparse are not a dict object -- trying to
use dict() as a default only causes it to fail. I ran into this when
trying to run server/conversion/cgivar_to_gff.py on the command line.

To fix this, the dict() default is replace with None. In addition,
tests which were checking the options data for values now first check
if options itself evaluates as "true".