Add conversion of deCODEme data
Also includes server/utils/autozip.py support for multi-file ZIP
archives, if one specifies a filename to open in the archive.

This is because deCODEme data is provided to users as a ZIP archive
with three files: 'deCODEme_info.txt', 'deCODEme_scan.csv', and
'deCODEme_sign.txt'. The second file contains the genotyping data.

The converter also supports receiving the deCODEme_scan.csv file
directly, rather than the whole archive -- this option is supported
because the deCODEme_info.txt contains the customer's username,
which they might not want to associated with the rest of the data.

Genome build is assumed to be build 36. Build information is
contained within deCODEme_info.txt, but that file is not examined,
or is not available if processing the deCODEme_scan.csv file directly.
As of 8/2012, we haven't seen build 37 data from deCODEme.